Editor component (#2)
- make editor component that is essentially textarea with highlight
capability (see discussion below)
- dummy api calls
- basic typescript types for audience, article, etc
- tweak button component
- fix Arjun's email
- make floating loading spinner component
- make "stat" component
- make "tool" higher-order-component for use on the home page and
article page
    - state interface
    - reducer for redux-like state with useReducer
- add title/source fields, author/date metadata, and action buttons,
shown/hidden based on logged in state and current page
    - implement word highlighting scores with editor component
- css tweaks
- various util funcs for string manipulation 

I spent several days experimenting with different methods to accomplish
the textarea with highlights.

The first method is just using a single `contenteditable` div. This is
what the current simplex app does. This comes with many downsides. 1)
Any time you programmatically make a change to the contents, the cursor
position resets, and preserving this is very convoluted because the
inner contents are arbitrary html, not plain text. 2) Every browser
implements `contenteditable` slightly differently, especially regarding
newlines, so the value of `innerText` is not consistent. I tried many
many things to normalize this across browsers, but I could not. It's too
complex of a problem space. 3) You have to manually intercept pastes and
force them to plain text. Not the most difficult thing, but a pain
still. 4) It is difficult to programmatically set (necessary when
filling example text, or loading an existing article) or render (like
using jsx to generate the highlights) the contents, again because of
cross browser `innerText` inconsistencies and react's resistance to
rendering raw html. All of these downsides add a lot of complexity and
fragility to edge cases and cross browser compatibility.

A second method I tried, to at least avoid the cursor jumping problem,
was to overlay two `contenteditable` divs. I still found the `innerText`
inconsistencies to be a deal breaker though.

A third method I tried, is using third party libraries mark.js to do the
highlighting. This was a pain to keep in sync with the react state of
the component, and unfortunately also had the problem of resetting the
cursor position after any change.

A fourth method I investigated was using some third party libraries to
do the whole thing:
https://github.com/bonafideduck/react-highlight-within-textarea
https://github.com/facebook/lexical

https://github.com/lonekorean/highlight-within-textarea/blob/master/jquery.highlight-within-textarea.js
The first of these seems to fit the needs, however it is based on a now
deprecated library called draft.js, and the "known issues" part of the
readme scares me (no accessibility support at all). Lexical seems to be
the "right" way to do this, but is still experimental and adds a lot of
weight and learning-curve for such a simple thing. Finally, I settled on
the last library, or rather my own simplified version of it.

It essentially overlays a normal `textarea` that handles the typing,
cursor position, selection, pasting, etc and shows the text, while an
underlay handles the highlight background (with their text color set to
invisible). Then, by figuring out a whole bunch of precise CSS styles,
you can make them wrap in (**theoretically**) the exact same way so that
the highlights stay "in-sync" with the typed text. Then you also have to
make sure their scrolls stay in sync. This is the cleanest solution
code-wise, but I'm not thrilled about this, because relying on the
browser to consistently text-wrap two separate elements in the exact
same way seems fragile. In fact, I was able to find a very specific case
where it fails:

<img width="638" alt="Screen Shot 2022-10-18 at 10 59 29 AM"
src="https://user-images.githubusercontent.com/8326331/196467250-78c212c7-b086-4d37-9ab3-14133e46d1b0.png">

However, I can only replicate this at one exact window width, on chrome,
in macos, at 100% zoom, with this font that I'm using. If any of those
change, it goes away. I also verified that
`jquery.highlight-within-textarea` (and indeed, several other
official/accepted libraries that use the same textarea method) fall prey
to this same problem at the same width/browser/zoom/font. So I am not
alone. I'm guessing this has to do with the fact that text wrapping
involves sub-pixel, floating point calculation of widths, and can go
wrong more easily on retina displays where 1 pixel =/= 1 pixel. I am
okay calling this a chrome bug.

I tried to fix this in every way possible, including looking up the most
obscure CSS properties I've heard of. I could not. This is an extreme
edge case I would say, and is worth the simplicity of the code and
accessibility (it's just a normal textarea input). If we decide this is
unacceptable, we could try one of the other libraries I listed above.